requirements.txt
torch
transformers
diffusers
accelerate
safetensors
fastapi
uvicorn
pyngrok
gradio
pydantic
python-dotenv
nest-asyncio
